refactor(react-components): do not use Link where we really just want the HashLink/anchor behavior
