+++ packages/react-components/src/atoms/Link.tsx
@@ -1,37 +1,22 @@
 import React, { ReactNode } from 'react';
-import css, { CSSObject, SerializedStyles } from '@emotion/css';
-import { HashLink } from 'react-router-hash-link';
+import css, { SerializedStyles } from '@emotion/css';
 
 import { fern, paper, pine } from '../colors';
-import { useHasRouter } from '../routing';
 import { ThemeVariant, defaultThemeVariant } from '../theme';
 import { getButtonStyles, getButtonChildren } from '../button';
-import { isInternalLink } from '../utils';
+import { Anchor } from '.';
 
 const styles = css({
-  outline: 'none',
-  textDecoration: 'none',
-  ':hover, :focus': {
-    textDecoration: 'none',
-  },
-  color: 'unset',
-  ':active': {
-    color: 'unset',
-  },
+  textDecoration: 'underline',
 });
-
 export const themeStyles: Record<ThemeVariant, SerializedStyles> = {
   light: css({ color: fern.rgb, ':active': { color: pine.rgb } }),
   grey: css({ color: fern.rgb, ':active': { color: pine.rgb } }),
   dark: css({ color: paper.rgb, ':active': { color: paper.rgb } }),
 };
-const underlineStyles = css({
-  textDecoration: 'underline',
-});
 
 interface NormalLinkProps {
-  readonly theme?: ThemeVariant | null;
-  readonly display?: CSSObject['display'];
+  readonly theme?: ThemeVariant;
 
   readonly buttonStyle?: undefined;
 
@@ -41,7 +26,6 @@ interface NormalLinkProps {
 }
 interface ButtonStyleLinkProps {
   readonly theme?: undefined;
-  readonly display?: undefined;
 
   readonly buttonStyle: true;
 
@@ -51,7 +35,6 @@ interface ButtonStyleLinkProps {
 }
 type LinkProps = {
   readonly children: ReactNode;
-  // hrefs may conditionally be undefined, but the prop is mandatory so it cannot be forgotton
   readonly href: string | undefined;
   readonly label?: string;
 } & (NormalLinkProps | ButtonStyleLinkProps);
@@ -61,7 +44,6 @@ const Link: React.FC<LinkProps> = ({
   label,
 
   theme = defaultThemeVariant,
-  display,
 
   buttonStyle = false,
   primary = false,
@@ -69,34 +51,13 @@ const Link: React.FC<LinkProps> = ({
   enabled = true,
 }) => {
   const linkStyles = buttonStyle
-    ? [styles, getButtonStyles({ primary, small, enabled, children })]
-    : [
-        styles,
-        theme && themeStyles[theme],
-        theme && underlineStyles,
-        { display },
-      ];
+    ? [getButtonStyles({ primary, small, enabled, children })]
+    : [styles, themeStyles[theme]];
   const linkChildren = buttonStyle ? getButtonChildren(children) : children;
-  const internal = enabled && href ? isInternalLink(href) : false;
-  if (useHasRouter() && href && internal) {
-    return (
-      <HashLink to={href} aria-label={label} css={linkStyles} smooth>
-        {linkChildren}
-      </HashLink>
-    );
-  }
   return (
-    // lint rule does not understand the conditionals
-    // eslint-disable-next-line react/jsx-no-target-blank
-    <a
-      href={(enabled && href) || undefined}
-      aria-label={label}
-      css={linkStyles}
-      target={internal ? undefined : '_blank'}
-      rel={internal ? undefined : 'noreferrer noopener'}
-    >
+    <Anchor href={href} enabled={enabled} aria-label={label} css={linkStyles}>
       {linkChildren}
-    </a>
+    </Anchor>
   );
 };
