+++ packages/react-components/src/atoms/__tests__/Link.test.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import { StaticRouter } from 'react-router-dom';
-import { render, fireEvent, waitFor } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
+import { render } from '@testing-library/react';
 import { findParentWithStyle } from '@asap-hub/dom-test-utils';
 
 import Link from '../Link';
@@ -45,28 +43,6 @@ describe('the dark theme', () => {
     );
     const { textDecoration } = getComputedStyle(getByRole('link'));
     expect(textDecoration).toBe('underline');
-  });
-});
-
-describe('no theme', () => {
-  it('applies no color', () => {
-    const { getByRole } = render(
-      <Link href="/" theme={null}>
-        text
-      </Link>,
-    );
-    const { color } = getComputedStyle(getByRole('link'));
-    expect(color).toBe('');
-  });
-
-  it('does not apply an underline', () => {
-    const { getByRole } = render(
-      <Link href="/" theme={null}>
-        text
-      </Link>,
-    );
-    const { textDecoration } = getComputedStyle(getByRole('link'));
-    expect(textDecoration).toBe('none');
   });
 });
 
@@ -150,125 +126,17 @@ describe('when button-styled', () => {
   });
 
   it('removes the href when disabled', () => {
-    const { getByText } = render(
+    const { getByText, rerender } = render(
+      <Link href="/" buttonStyle>
+        text
+      </Link>,
+    );
+    expect(getByText('text').closest('a')).toHaveAttribute('href');
+    rerender(
       <Link href="/" buttonStyle enabled={false}>
         text
       </Link>,
     );
-    expect(getByText('text')).not.toHaveAttribute('href');
-  });
-});
-
-describe.each`
-  contextDescription    | wrapper
-  ${'with a router'}    | ${StaticRouter}
-  ${'without a router'} | ${undefined}
-`('$contextDescription', ({ wrapper }) => {
-  describe.each`
-    linkDescription    | href
-    ${'external link'} | ${'https://parkinsonsroadmap.org/'}
-    ${'internal link'} | ${'/'}
-  `("for an $linkDescription to '$href'", ({ href }) => {
-    it('applies the href to the anchor', () => {
-      const { getByRole } = render(<Link href={href}>text</Link>, {
-        wrapper,
-      });
-      const anchor = getByRole('link') as HTMLAnchorElement;
-      expect(new URL(anchor.href, window.location.href).href).toBe(
-        new URL(href, window.location.href).href,
-      );
-    });
-  });
-
-  it('renders an inactive link without an href', () => {
-    const { getByText } = render(<Link href={undefined}>text</Link>, {
-      wrapper,
-    });
-    expect(getByText('text', { selector: 'a' })).not.toHaveAttribute('href');
-  });
-
-  it('renders an inactive link with an empty href', () => {
-    const { getByText } = render(<Link href="">text</Link>, {
-      wrapper,
-    });
-    expect(getByText('text', { selector: 'a' })).not.toHaveAttribute('href');
-  });
-});
-
-describe('for an external link', () => {
-  it('sets the anchor target to open in a new page', () => {
-    const { getByRole } = render(
-      <Link href="https://parkinsonsroadmap.org/">text</Link>,
-    );
-    const { target } = getByRole('link') as HTMLAnchorElement;
-    expect(target).toBe('_blank');
-  });
-
-  it('secures the link against third parties', () => {
-    const { getByRole } = render(
-      <Link href="https://parkinsonsroadmap.org/">text</Link>,
-    );
-    const { relList } = getByRole('link') as HTMLAnchorElement;
-    expect(relList).toContain('noreferrer');
-    expect(relList).toContain('noopener');
-  });
-
-  it('triggers a full page navigation on click', () => {
-    const { getByRole } = render(
-      <Link href="https://parkinsonsroadmap.org/">text</Link>,
-    );
-    const anchor = getByRole('link') as HTMLAnchorElement;
-    expect(fireEvent.click(anchor)).toBe(true);
-  });
-});
-
-describe.each`
-  description           | wrapper
-  ${'with a router'}    | ${StaticRouter}
-  ${'without a router'} | ${undefined}
-`('for an internal link $description to /', ({ wrapper }) => {
-  it('does not set the anchor target', () => {
-    const { getByRole } = render(<Link href="/">text</Link>, { wrapper });
-    const { target } = getByRole('link') as HTMLAnchorElement;
-    expect(target).toBe('');
-  });
-
-  it('does not secure the link against third parties', () => {
-    const { getByRole } = render(<Link href="/">text</Link>, { wrapper });
-    const { relList } = getByRole('link') as HTMLAnchorElement;
-    expect(relList).not.toContain('noreferrer');
-    expect(relList).not.toContain('noopener');
-  });
-});
-
-describe('for an internal link with a router', () => {
-  it('does not trigger a full page navigation on click', () => {
-    const { getByRole } = render(
-      <Link
-        href={`${window.location.protocol}//${window.location.host}/page?query#fragment`}
-      >
-        text
-      </Link>,
-      { wrapper: StaticRouter },
-    );
-    const anchor = getByRole('link') as HTMLAnchorElement;
-    expect(fireEvent.click(anchor)).toBe(false);
-  });
-
-  it('smoothly scrolls the anchor referenced by the fragment into view', async () => {
-    const { getByRole } = render(
-      <>
-        <Link href={`#fragment`}>text</Link>
-        <main id="fragment">text</main>
-      </>,
-      { wrapper: StaticRouter },
-    );
-    const main = getByRole('main');
-    const spyScrollIntoView = jest.spyOn(main, 'scrollIntoView');
-
-    userEvent.click(getByRole('link'));
-    await waitFor(() =>
-      expect(spyScrollIntoView).toHaveBeenCalledWith({ behavior: 'smooth' }),
-    );
+    expect(getByText('text').closest('a')).not.toHaveAttribute('href');
   });
 });
