+++ packages/react-components/src/atoms/index.ts
@@ -1,3 +1,4 @@
+export { default as Anchor } from './Anchor';
 export { default as Avatar } from './Avatar';
 export { default as Button } from './Button';
 export { default as Caption } from './Caption';
